% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_metrics.R
\name{d_dist}
\alias{d_dist}
\title{Disturbance impact}
\usage{
d_dist(ts, ys, tpert = 0, ts_pre = -1)
}
\arguments{
\item{ts}{Vector containing the times (same size as ys). Perturbation assumed to happen at 0}

\item{ys}{Vector containing the values (same size as ts)}

\item{tpert}{Time of the perturbation. Default set to 0 yr}

\item{ts_pre}{Sampling times for estimating Vpre. Default set to -1}
}
\value{
the disturbance impact
}
\description{
Disturbance impact
}
