% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fire.R
\name{extract_date_fire}
\alias{extract_date_fire}
\title{Extract date from CCI fire image name}
\usage{
extract_date_fire(x)
}
\arguments{
\item{x}{image name}
}
\value{
date
}
\description{
Extract date from CCI fire image name
}
