% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fire.R
\name{getFireDate}
\alias{getFireDate}
\title{Extract the dates of fire events}
\usage{
getFireDate(ts_conf, ts_doy, dts, conf_level)
}
\arguments{
\item{ts_conf}{time series of the fire confidence level (vector of observations)}

\item{ts_doy}{time series of the fire day of year (vector of observations)}

\item{dts}{dates (vector of Date objects)}

\item{conf_level}{minimum fire confidence level to consider a date as fire event}
}
\value{
dates of fire events
}
\description{
Extract the dates of fire events
}
