% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_clim.R
\name{getMonthlyMeanClim}
\alias{getMonthlyMeanClim}
\title{Get the mean climate variable for each month using data after a predefined start date}
\usage{
getMonthlyMeanClim(br, startdt)
}
\arguments{
\item{br}{brick of CRU climate data}

\item{startdt}{start date of the data that should be processed}
}
\value{
mean monthly climate variable
}
\description{
Get the mean climate variable for each month using data after a predefined start date
}
