% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_ts.R
\name{getRegularTS}
\alias{getRegularTS}
\title{Get regular time series}
\usage{
getRegularTS(dts, obs, fun = "mean", resol = "monthly")
}
\arguments{
\item{dts}{vector of dates (formatted as strings YYYYMMDD)}

\item{obs}{vector of observations}

\item{fun}{function used to aggregate observations}

\item{resol}{desired resolution of the regular time series}
}
\value{
regular time series
}
\description{
Convert irregular time series with missing values to a regular one
}
