% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_ts.R
\name{mmax}
\alias{mmax}
\title{Helper function for the toRegularTS function}
\usage{
mmax(x)
}
\arguments{
\item{x}{vector of observations}
}
\value{
the maximum value of the vector
}
\description{
Helper function for the toRegularTS function
}
