% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_aux}
\alias{plot_aux}
\title{Plot recovery metrics}
\usage{
plot_aux(
  dts,
  obsi,
  main,
  lossdt_start,
  lossdt_end,
  predt_start,
  distdt_end,
  postdt_start,
  postdt_end,
  trend,
  met_rri,
  met_r80p,
  met_yryr,
  ...
)
}
\arguments{
\item{dts}{vector of dates}

\item{obsi}{time series - vector of observations}

\item{main}{title of the plot (string)}

\item{lossdt_start}{start uncertainty window of disturbance date}

\item{lossdt_end}{end uncertainty window of disturbance date}

\item{trend}{fitted trend component - vector of observations}

\item{met_rri}{derived RRI metric}

\item{met_r80p}{derived R80p metric}

\item{met_yryr}{derived YrYr metric}

\item{lossdt_start1}{start date for time window of accepted breaks}

\item{lossdt_end1}{end date for time window of accepted breaks}
}
\value{
plot
}
\description{
Plot recovery metrics
}
