% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot_BRT_hist}
\alias{plot_BRT_hist}
\title{Plot histogram and marginal effects}
\usage{
plot_BRT_hist(mod_brt, dat_covar, names_covars, names_covars_short)
}
\arguments{
\item{mod_brt}{Boosted regression tree model}

\item{dat_covar}{dataframe of covariate values}

\item{names_covars}{names of the covariates (as used in BRT and the dataframe)}

\item{names_covars_short}{the covariate names used for plotting (order should be the same as names_covars)}
}
\value{
plot
}
\description{
Plot histogram and marginal effects
}
