% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_covarData}
\alias{tidy_covarData}
\title{Tidy covariate data downloaded from GEE}
\usage{
tidy_covarData(x, col_names)
}
\arguments{
\item{x}{the input data}

\item{col_names}{column names associated with the input data}
}
\value{

}
\description{
Tidy covariate data downloaded from GEE
}
