% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy GEE data}
\usage{
tidy_data(x, col_names)
}
\arguments{
\item{x}{input data}

\item{col_names}{column names associated with the input data}
}
\value{

}
\description{
Tidy GEE data
}
