% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fire.R
\name{tidy_fire_date}
\alias{tidy_fire_date}
\title{Extract a date frame of fire doy and fire confidence level from data downloaded via GEE}
\usage{
tidy_fire_date(dat_fire)
}
\arguments{
\item{dat_fire}{dataframe of fire related data}
}
\value{
date frame of fire doy and fire confidence level
}
\description{
Extract a date frame of fire doy and fire confidence level from data downloaded via GEE
}
