% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_ts.R
\name{toRegularTS}
\alias{toRegularTS}
\title{Generate regular time series}
\usage{
toRegularTS(tsi, dts, fun, resol)
}
\arguments{
\item{tsi}{vector of observations}

\item{dts}{vector of dates}

\item{fun}{function to aggregate ('max' or 'mean')}

\item{resol}{temporal resolution ('monthly', 'daily', or 'quart')}
}
\value{
regular time series
}
\description{
Generate regular time series
}
