% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_metrics.R
\name{V_pre}
\alias{V_pre}
\title{Pre-disturbance state}
\usage{
V_pre(ts, ys, ts_pre = c(-1, -2))
}
\arguments{
\item{ts}{Vector containing the times (same size as ys). Perturbation assumed to happen at 0}

\item{ys}{Vector containing the values (same size as ts)}

\item{ts_pre}{Sampling times for estimating Vpre. Default set to -1}
}
\value{
the pre-disturbance state
}
\description{
Pre-disturbance state
}
