% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_metrics.R
\name{calcFrazier}
\alias{calcFrazier}
\title{Calculate recovery metrics}
\usage{
calcFrazier(
  tsio,
  tdist,
  obspyr,
  nPre,
  nDist,
  nPost,
  nPostStart,
  nDelta,
  nDeltaStart
)
}
\arguments{
\item{tsio}{vector of observations (time series with a fixed observation frequency)}

\item{tdist}{observation number of disturbance, indicating the timing of the disturbance}

\item{obspyr}{number of observations per year}

\item{nPre}{number of years prior to the disturbance used to calculate the pre-disturbance value}

\item{nDist}{number of years used to quantify the state during the disturbance}

\item{nPost}{number of years used to calculate the post-disturbance value}

\item{nPostStart}{indicates the start of the post-disturbance period (expressed as number of years after the disturbance)}

\item{nDelta}{number of years used to calculate the post-disturbance value for the YrYr metric}

\item{nDeltaStart}{indicates the start of the post-disturbance period for the YrYr metric (expressed as number of years after the disturbance)}
}
\value{
a list containing the RRI recovery indicator, R80p recovery indicator, YrYr recovery indicator, disturbance impact, and pre-disturbance state
}
\description{
Calculate recovery metrics from a time series with known disturbance date. The calcFrazier function derives the RRI, R80P and YrYr recovery indicators,
defined by Frazier et al. (2018). The indicators are originally developed for annual long-term time series of optical vegetation indices.
Yet, in order to be able to derive the indicators as well for dense and/or short time series, a modified version is suggested.
Here, the user can define the time period before, during and after the disturbance that is used to derive the indicators.
To reduce the interference of the seasonal pattern of dense time series, the chosen time period should cover blocks of n years.
(Frazier, R. J., Coops, N. C., Wulder, M. A., Hermosilla, T., & White, J. C. (2018). Analyzing spatial and temporal variability in short-term rates
of post-fire vegetation return from Landsat time series. Remote Sensing of Environment, 205, 32-45.)
}
