% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_metrics.R
\name{calcRecMetrics}
\alias{calcRecMetrics}
\title{Calculate recovery metrics}
\usage{
calcRecMetrics(
  tsi,
  tdist,
  obspyr,
  nPre,
  nDist,
  nPost,
  nPostStart,
  nDelta,
  nDeltaStart,
  tbp,
  maxBreak,
  chkBrk = F,
  timeThres
)
}
\arguments{
\item{timeThres}{}
}
\value{

}
\description{
Calculate recovery metrics
}
