% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_ts.R
\name{getRegularTS}
\alias{getRegularTS}
\title{Get regular time series}
\usage{
getRegularTS(dts, obs)
}
\arguments{
\item{dts}{vector of dates (formatted as strings YYYYMMDD)}

\item{obs}{vector of observations}
}
\value{
regular time series
}
\description{
Convert irregular time series with missing values to a regular one
}
