% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_aux}
\alias{plot_aux}
\title{Title}
\usage{
plot_aux(
  dts,
  obsi,
  main,
  lossdt_start,
  lossdt_end,
  lossdt_start1,
  lossdt_end1,
  trend,
  met_rri,
  met_r80p,
  met_yryr
)
}
\arguments{
\item{met_yryr}{}
}
\value{

}
\description{
Title
}
